--- src/bundles.rs
@@ -1,2 +1,0 @@
-pub mod physics;
-pub mod player;
