+++ src/data.rs
@@ -2,6 +2,7 @@ use bevy::prelude::Plugin;
 
 pub mod player;
 pub mod tilemap;
+pub mod physics;
 
 pub struct ComponentsPlugin;
 impl Plugin for ComponentsPlugin {
