+++ src/data/player.rs
@@ -1,4 +1,7 @@
 use bevy::prelude::*;
+use bevy_rapier2d::prelude::*;
+
+use super::physics::RigidBodyBundle;
 
 pub fn register(app: &mut bevy::prelude::App) {
     app
@@ -47,3 +50,32 @@ pub struct PlayerFoot;
 
 #[derive(Component, Default, Reflect)]
 pub struct PlayerHead;
+
+#[derive(Bundle, Default)]
+pub struct PlayerBundle {
+    pub sprite: SpriteSheetBundle,
+    pub physic: RigidBodyBundle,
+    pub player: Player,
+    pub jump: PlayerJump,
+    pub info: PlayerInfo,
+    pub movement: PlayerMovement,
+    pub name: Name,
+}
+
+#[derive(Bundle, Default)]
+pub struct PlayerFootBundle {
+    pub transform: TransformBundle,
+    pub collider: Collider,
+    pub sensor: Sensor,
+    pub foot_comp: PlayerFoot,
+    pub name: Name
+}
+
+#[derive(Bundle, Default)]
+pub struct PlayerHeadBundle {
+    pub transform: TransformBundle,
+    pub collider: Collider,
+    pub sensor: Sensor,
+    pub head_comp: PlayerHead,
+    pub name: Name
+}
