+++ src/plugins/player.rs
@@ -1,11 +1,10 @@
 mod jump;
 mod movement;
+pub mod spawn;
 
 use bevy::prelude::*;
 
-use crate::prefabs::player::spawn_player;
-
-use self::{jump::{handle_jump, check_if_grounded, check_if_head_bump}, movement::{handle_movement, apply_accel_when_land}};
+use self::{jump::{handle_jump, check_if_grounded, check_if_head_bump}, movement::{handle_movement, apply_accel_when_land}, spawn::spawn_player};
 
 pub enum PlayerEvent {
     Airborne(Entity),
