+++ src/plugins/tilemap.rs
@@ -1,6 +1,6 @@
 use std::collections::{HashMap, HashSet};
 
-use crate::{entities::{TerrainBundle, TrapBundle, WebStickerBundle, Terrain, WebSticker, Trap}, components::tilemap::Platform};
+use crate::{entities::{TerrainBundle, TrapBundle, WebStickerBundle, Terrain, WebSticker, Trap}, data::tilemap::Platform};
 use bevy::prelude::*;
 use bevy_ecs_ldtk::{prelude::*, GridCoords};
 use bevy_rapier2d::prelude::*;
