--- src/prefabs.rs
@@ -1,3 +1,0 @@
-use bevy::prelude::Entity;
-
-pub mod player;
